requirements.txt
yfinance==0.2.28
pandas==2.2.2
numpy==1.24.3
streamlit==1.37.0
plotly==5.17.0
scipy==1.11.4
seaborn==0.13.2
matplotlib==3.7.5
python-telegram-bot==20.7
requests==2.31.0
pyyaml==6.0.1